backend/requirements.txt
flask==3.1.1
flask-cors==6.0.0
flask-restful==0.3.10
marshmallow==4.0.0
python-dotenv==1.1.0
Werkzeug==3.1.3